requirements.txt
streamlit==1.40.2
Pillow==9.5.0
torch==2.5.1
torchvision==0.20.1
requests
uvicorn
fastapi